BUG: Removed warning messages on Win7 with VS12
 - Removed all warning messages related to data-casting
   on Windows7 system with Visual Studio 2013

+++ kwsCheckEmptyLines.cxx
@@ -26,7 +26,7 @@ bool Parser::CheckEmptyLines(unsigned long max, bool checkEndOfFile)
   delete [] val;
 
   bool hasError = false;
-  unsigned long total = m_Buffer.size();
+  unsigned long total = static_cast<unsigned long>(m_Buffer.size());
   unsigned long i = 0;
   unsigned long j = 1;
   unsigned long empty = 0;
@@ -79,7 +79,7 @@ bool Parser::CheckEmptyLines(unsigned long max, bool checkEndOfFile)
         }
       }
     j++;
-    i += line.length()+1;
+    i += static_cast<unsigned long>(line.length())+1;
     }
   return !hasError;
 }
