+++ kwsCheckVariablePerLine.cxx
@@ -44,7 +44,7 @@ bool Parser::CheckVariablePerLine(unsigned long max)
       {
       // Check that this is the first word
       bool firstWord = false;
-      long int pos=posType;
+      long int pos=static_cast<long int>(posType);
       pos--;
       while((pos>0) && (m_BufferNoComment[pos]==' '))
         {
@@ -60,15 +60,15 @@ bool Parser::CheckVariablePerLine(unsigned long max)
         std::string line = this->GetLine(this->GetLineNumber(posType,true)-1);
 
         // Check if we have any comments
-        int poscom = line.find("//",0);
+        int poscom = static_cast<int>(line.find("//",0));
         if(poscom != -1)
           {
           line = line.substr(0,poscom);
           }
-        poscom = line.find("/*",0);
+        poscom = static_cast<int>(line.find("/*",0));
         while(poscom != -1)
           {
-          int poscomend = line.find("*/",0);
+          int poscomend = static_cast<int>(line.find("*/",0));
           if(poscomend == -1)
             {
             line = line.substr(0,poscom);
@@ -77,7 +77,7 @@ bool Parser::CheckVariablePerLine(unsigned long max)
 
           std::string line_temp = line.substr(0,poscom);
           line = line_temp+line.substr(poscomend+2,line.size()-poscomend-2);
-          poscom = line.find("/*",poscom+1);
+          poscom = static_cast<int>(line.find("/*",poscom+1));
           }
 
         // If we have any '(' in the line we stop
@@ -85,7 +85,7 @@ bool Parser::CheckVariablePerLine(unsigned long max)
           {
           // This is a very simple check we count the number of comas
           unsigned int vars = 1;
-          pos = line.find(',',0);
+          pos = static_cast<long int>(line.find(',',0));
           while(pos!=-1)
             {
             // Check that we are not initializing an array
@@ -96,7 +96,7 @@ bool Parser::CheckVariablePerLine(unsigned long max)
               // Ok we have the opening
               if(line[openCurly] == '{')
                 {
-                long int posClosing = this->FindClosingChar('{','}',openCurly,false,line);
+                long int posClosing = static_cast<long int>(this->FindClosingChar('{','}',openCurly,false,line));
                 if(posClosing == -1
                   || pos<posClosing)
                   {
@@ -125,7 +125,7 @@ bool Parser::CheckVariablePerLine(unsigned long max)
               {
               vars++;
               }
-            pos = line.find(',',pos+1);
+            pos = static_cast<long int>(line.find(',',pos+1));
             }
 
           if(vars > max)
